// src/app/model/actividad.ts
export interface Actividad {

  actSecuencia: number;
  tdcTdEpl: string;
  eplNd: number;
  actFecha: string;
  actActividad: string;
  actDescripcion: string;
  actTiempo: any;
  actUsuario: string;
  actOperacion: string;
  radicacion: number;
  actDependencia: string;
  actProyecto: string;
  actTipo: string;

}