// src/app/model/consultaActividad.ts
export interface ConsultaActividad {

  actSecuencia: number;
  tdcTdEpl: string;
  eplNd: number;
  actFecha: string;
  actActividad: string;
  actDescripcion: string;
  actTiempo: number;
  actUsuario: string;
  actOperacion: string;
  radicacionTexto: number;
  actDependencia: string;
  actProyecto: string;
  actTipo: string;

}